moved BigQuerySourceConfig
